Move WidgetFamily environment key to separate file
Refactored the custom WidgetFamily environment key and EnvironmentValues extension into a new WidgetFamily+AppPreview.swift file for better modularity. Updated WidgetBasicContainerView to use the new location and wrapped .environment calls with #if !WIDGET_EXTENSION for clarity.
