+++ Sources/Extensions/Widgets/Common/WidgetBasicContainerView.swift
@@ -34,21 +34,6 @@ struct WidgetBasicContainerView: View {
     }
 }
 
-// Only define custom widgetFamily environment when not compiling for the widget extension,
-// as WidgetKit already provides the native \.widgetFamily environment variable.
-#if !WIDGET_EXTENSION
-extension WidgetFamily: @retroactive EnvironmentKey {
-    public static var defaultValue: WidgetFamily = .systemMedium
-}
-
-extension EnvironmentValues {
-    var widgetFamily: WidgetFamily {
-        get { self[WidgetFamily.self] }
-        set { self[WidgetFamily.self] = newValue }
-    }
-}
-#endif
-
 @available(iOS 18, *)
 struct WidgetBasicContainerView_Previews: PreviewProvider {
     struct WidgetBasicContainerViewPreviewData {
@@ -74,7 +59,9 @@ struct WidgetBasicContainerView_Previews: PreviewProvider {
                 familySize: .systemSmall
             )
             .previewContext(WidgetPreviewContext(family: WidgetFamily.systemSmall))
-            .environment(\.widgetFamily, .systemSmall)
+            #if !WIDGET_EXTENSION
+                .environment(\.widgetFamily, .systemSmall)
+            #endif
         }
 
     static var systemMediumConfigurations: SnapshottablePreviewConfigurations<WidgetBasicContainerViewPreviewData> =
@@ -88,7 +75,9 @@ struct WidgetBasicContainerView_Previews: PreviewProvider {
                 familySize: .systemMedium
             )
             .previewContext(WidgetPreviewContext(family: WidgetFamily.systemMedium))
-            .environment(\.widgetFamily, .systemMedium)
+            #if !WIDGET_EXTENSION
+                .environment(\.widgetFamily, .systemMedium)
+            #endif
         }
 
     static var systemLargeConfigurations: SnapshottablePreviewConfigurations<WidgetBasicContainerViewPreviewData> =
@@ -102,7 +91,9 @@ struct WidgetBasicContainerView_Previews: PreviewProvider {
                 familySize: .systemLarge
             )
             .previewContext(WidgetPreviewContext(family: WidgetFamily.systemLarge))
-            .environment(\.widgetFamily, .systemLarge)
+            #if !WIDGET_EXTENSION
+                .environment(\.widgetFamily, .systemLarge)
+            #endif
         }
 
     private static func maxTiles(for familySize: WidgetFamily) -> Int {
